// src/app/models/Cars.ts
export class Cars {
    CarsId: number;
    Ccy: string;
    Title: string;
    ShortDescrip: string;
    Descrip: string;
    ExtraImages: string;
    Date: Date;
    ImageUrl: string;
    Lang: string;
    Color: number;
    Category: number;
    Transmission: number;
    NumberOfDoors: number;
    NumberOfSeats: number;
    Engine: number;
    Price: number;
    FuelType: number;
    Mileage: number;
    DriveWheels: number;
}